Added networking
Added VERY simple netwroking. No replication.
One player can host and the other can join.

Also:
Changed default level to Main Menu
created a new gameinstance and set it as the default.
Fixed an error about the flying pawn having an empty mesh. It was because it had a sphere as its root with no mesh. deleted it and created a scene node and set it as root.